chore: Update website links and images
The package.json file was updated to change the homepage link to the new website URL. Additionally, the build-gh script was modified to use the correct public URL for deployment.

In the App.jsx file, the siteProps object was updated with new information for the web designer and developer. The primaryColor variable was also modified.

The About.jsx file was updated to change the image and imageAltText variables to reflect the new background image and description.

The Home.jsx file was updated to change the image and imageAltText variables to reflect the new background image.

The Portfolio.jsx file was updated to change the image and imageAltText variables to reflect the new background image. The projectList array was also modified to include new projects.

commit message goes here